requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pydantic2
PyJWT
cryptography